:recycle: refactor oracle executor into modular command handlers
- Reduced 1,135-line OracleActionExecutor monolith to a 127-line thin dispatcher.
- Extracted slash commands and AI logic into 9 specialized executors in packages/oracle-engine/src/executors/.
- Implemented Command + Event + DI hybrid architecture using @codex/events.
- Added isExecuting guards to prevent race conditions during AI streaming.
- Decoupled UI side effects (notifications) via global event listeners.
- Achieved 82.5% unit test coverage for the new package.
- Preserved guest mode restrictions and Fog of War visibility checks.
