+++ packages/oracle-engine/src/events.ts
@@ -1,7 +1,13 @@
 import type { AppEventDefinition } from "@codex/events";
+import type { OracleIntent, DiscoveryProposal } from "./types";
 
 export const ORACLE_EVENTS = {
   UNDO_PERFORMED: "ORACLE:UNDO_PERFORMED",
+  COMMAND_STARTED: "ORACLE:COMMAND_STARTED",
+  COMMAND_COMPLETED: "ORACLE:COMMAND_COMPLETED",
+  COMMAND_FAILED: "ORACLE:COMMAND_FAILED",
+  ENTITY_DISCOVERED: "ORACLE:ENTITY_DISCOVERED",
+  ENTITY_CREATED: "ORACLE:ENTITY_CREATED",
 } as const;
 
 export type OracleEventType =
@@ -12,6 +18,26 @@ declare module "@codex/events" {
     "ORACLE:UNDO_PERFORMED": AppEventDefinition<
       "oracle",
       { messageId: string }
+    >;
+    "ORACLE:COMMAND_STARTED": AppEventDefinition<
+      "oracle",
+      { intent: OracleIntent }
+    >;
+    "ORACLE:COMMAND_COMPLETED": AppEventDefinition<
+      "oracle",
+      { intent: OracleIntent; result?: any }
+    >;
+    "ORACLE:COMMAND_FAILED": AppEventDefinition<
+      "oracle",
+      { intent: OracleIntent; error: string }
+    >;
+    "ORACLE:ENTITY_DISCOVERED": AppEventDefinition<
+      "oracle",
+      { proposal: DiscoveryProposal }
+    >;
+    "ORACLE:ENTITY_CREATED": AppEventDefinition<
+      "oracle",
+      { entityId: string; title: string }
     >;
   }
 }
