+++ packages/oracle-engine/src/executors/dice-executor.ts
@@ -3,40 +3,67 @@ import type {
   OracleIntent,
   OracleExecutionContext,
 } from "../types";
+import { BaseExecutor } from "./base-executor";
+import { ORACLE_EVENTS } from "../events";
 
-export class DiceExecutor implements OracleCommandExecutor {
+export class DiceExecutor
+  extends BaseExecutor
+  implements OracleCommandExecutor
+{
   async execute(
     intent: OracleIntent,
     context: OracleExecutionContext,
   ): Promise<void> {
-    const formula = intent.formula;
-    if (!formula) {
-      await context.chatHistory.addMessage({
-        id: crypto.randomUUID(),
-        role: "system",
-        content: "❌ Please specify a roll formula (e.g. /roll 1d20).",
+    await this.executeWithStack(intent, context, async () => {
+      await this.emit(context, {
+        type: ORACLE_EVENTS.COMMAND_STARTED,
+        payload: { intent },
       });
-      return;
-    }
 
-    try {
-      const command = context.diceParser.parse(formula);
-      const result = context.diceEngine.execute(command);
-      await context.diceHistory.addResult(result, "chat");
+      const formula = intent.formula;
+      if (!formula) {
+        const error = "❌ Please specify a roll formula (e.g. /roll 1d20).";
+        await context.chatHistory.addMessage({
+          id: crypto.randomUUID(),
+          role: "system",
+          content: error,
+        });
+        await this.emit(context, {
+          type: ORACLE_EVENTS.COMMAND_FAILED,
+          payload: { intent, error },
+        });
+        return;
+      }
 
-      await context.chatHistory.addMessage({
-        id: crypto.randomUUID(),
-        role: "system",
-        content: "",
-        type: "roll",
-        rollResult: result,
-      });
-    } catch (err: any) {
-      await context.chatHistory.addMessage({
-        id: crypto.randomUUID(),
-        role: "system",
-        content: `❌ Roll failed: ${err.message}`,
-      });
-    }
+      try {
+        const command = context.diceParser.parse(formula);
+        const result = context.diceEngine.execute(command);
+        await context.diceHistory.addResult(result, "chat");
+
+        await context.chatHistory.addMessage({
+          id: crypto.randomUUID(),
+          role: "system",
+          content: "",
+          type: "roll",
+          rollResult: result,
+        });
+
+        await this.emit(context, {
+          type: ORACLE_EVENTS.COMMAND_COMPLETED,
+          payload: { intent, result },
+        });
+      } catch (err: any) {
+        const error = err.message || "Unknown roll error";
+        await context.chatHistory.addMessage({
+          id: crypto.randomUUID(),
+          role: "system",
+          content: `❌ Roll failed: ${error}`,
+        });
+        await this.emit(context, {
+          type: ORACLE_EVENTS.COMMAND_FAILED,
+          payload: { intent, error },
+        });
+      }
+    });
   }
 }
